initial commit
basic functionality. cells change color when clicked. available moves highlight when hovering on a cell.  Using resources for the data model. I want to separate the presentation from the data as much as possible. Not sure if that is great form, I am trying my best to apply object oriented approaches.